Partially implement prototype path generator
A partial implementation of the prototype path generator is implemented
and found to be too slow. The bottleneck is believed to be the frequent
disk access for pre-calculated shortest-path distance. In light of this
and the fact that a story's context concepts remain unchanged between
evaluations, the relevance score for each concept of the knowledge base
can be calculated, ahead of time or upon initialization, for each story
(evaluator would be bounded to story instance). The resulting relevance
scores can be stored in memory, relieving the disk access bottleneck.